+++ sqldelight-gradle-plugin/src/test/integration-postgresql/src/test/kotlin/app/cash/sqldelight/postgresql/integration/PostgreSqlTest.kt
@@ -77,7 +77,8 @@ class PostgreSqlTest {
 
   @Before fun before() {
     driver.execute(null, "SET timezone TO UTC", 0)
-    driver.execute(null, "CREATE EXTENSION IF NOT EXISTS postgis", 0)
+    driver.execute(null, "CREATE EXTENSION IF NOT EXISTS postgis ", 0)
+    driver.execute(null, "CREATE EXTENSION IF NOT EXISTS ltree ", 0)
     MyDatabase.Schema.create(driver)
   }
 
@@ -1577,6 +1578,38 @@ class PostgreSqlTest {
       assertThat(it.cmin).isNotNull()
       assertThat(it.cmax).isNotNull()
       assertThat(it.ctid).isNotNull()
+    }
+  }
+
+  @Test
+  fun testLtreeContains() {
+    database.ltreeQueries.insertPath("Top")
+    database.ltreeQueries.insertPath("Top.Science")
+    database.ltreeQueries.insertPath("Top.Science.Astronomy")
+    database.ltreeQueries.insertPath("Top.Science.Astronomy.Astrophysics")
+    database.ltreeQueries.insertPath("Top.Science.Astronomy.Cosmology")
+    database.ltreeQueries.selectPathContains().executeAsList().let {
+      assertThat(it).containsExactly(
+        "Top.Science",
+        "Top.Science.Astronomy",
+        "Top.Science.Astronomy.Astrophysics",
+        "Top.Science.Astronomy.Cosmology",
+      )
+    }
+  }
+
+  @Test
+  fun testLtreeExists() {
+    database.ltreeQueries.insertPath("Top")
+    database.ltreeQueries.insertPath("Top.Science")
+    database.ltreeQueries.insertPath("Top.Science.Astronomy")
+    database.ltreeQueries.insertPath("Top.Science.Astronomy.Astrophysics")
+    database.ltreeQueries.insertPath("Top.Science.Astronomy.Cosmology")
+    database.ltreeQueries.selectPathExists("{Top.Science.Astronomy.Astrophysics, Top.Science.Astronomy.Cosmology}").executeAsList().let {
+      assertThat(it).containsExactly(
+        "Top.Science.Space.Astronomy.Astrophysics",
+        "Top.Science.Space.Astronomy.Cosmology",
+      )
     }
   }
 }