+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - [PostgreSQL Dialect] Add query support for implicitly defined System Columns (#5834 by @griffio)
 - [PostgreSQL Dialect] Add basic Array literal support (#5997 by @griffio)
+- [PostgreSQL Dialect] Add basic LTREE support (#5880 by @yesitskev @griffio)
 
 ### Changed
 - [PostgreSQL Dialect] Change arrayIntermediateType visibility to public (#5835 by @griffio)